ADVISOR-2390
This ticket is in regards to global filters not working in the systems table in advisor.

The issue was some params stuff we were sending the api.
To verify that the global filter works or replicate steps,
navigate to systems page in advisor, select a global filter (like Micorsoft SQL)
and count the amount of systems available to your account.

Global filters work fine in inventory, so head over to inventory with the same global filter selected, and you should see the same amount of systems seen previously in advisor<systems.