fix(RHINENG-9378): Fix recommendations bulk select
The issue described in the card has been fix by some other PR:

Go to recommendation's details page with at least 11 systems so it's is possible to have at least 2 pages in the table
Select All systems via bulk selector
Go to the next page of the table

Before, once you went to the next page, you would see that none of the systems on subsequent pages were selected. And clicking a system to select it would suddently select every system on the page. When I began working on this card, that issue was already fixed.

But, this PR fixes a couple of other issues:

When you choose Select All from the bulk select, it disables the bulk select box and the Spinner shows up properly
When you choose Select All from the bulk select, and the selection completes, you will see the checkbox in the bulk select next to the total number of systems selected