feat(RHINENG-3020) Updgrade to pdf-generator v2
This PR leverages the use of pdf-generator service to download the
executive report on the Advisor -> Recommendations page.

To test, go to Advisor -> Recommendations and click the button link in
the top right corner of the page to "Download Executive Report".

The new pdf-generator service is gated behind a new stage feature flag,
`advisor.pdf_generator`. By default, the feature flag is disabled; and
with the flag disabled, the export should act the same as it usually
does.

With the flag enabled the export will fail in this PR, which is
expected. It needs to be tested merged in stage, so we have the feature
flag to keep our release unblocked.