feat(RHINENG-17046): use new pdf-generator for systems table
This PR moves the systems table pdf export to the pdf-generator service
and updates the pdf template.

To run this locally, there is a multi-step process:

In the config/webpack.config.js file, add the following under the routes object:
    '/apps/advisor/': {
       host: 'http://localhost:8003',
     },
     '/api/crc-pdf-generator/': {
       host: 'http://localhost:8000',
     },

Then in a terminal window:
[insights-chrome]$ npm run dev

In a second terminal window, pull down pdf-generator make sure you're using node version 20. and run:
[pdf-generator]$ PROXY_AGENT=http://squid.corp.redhat.com:3128 ASSETS_HOST=https://localhost:1337 API_HOST=https://console.stage.redhat.com npm run start:server

In a third terminal window, navigate to pdf-generator and make sure you're using node version 20. Run:
[pdf-generator]$ podman-compose up

Finally, in a fourth terminal window, navigate to insights-advisor-frontend and run:
[insights-advisor-frontend]$ npm run static

Navigate to Advisor -> Systems
Click the Download icon in the table toolbar
Click Export to PDF

This should render a PDF that looks similar to the previous PDF in layout, but different in design. The data should be the same as the previous PDF.

This feature should be hidden behind a feature flag.

You should be able to filter the data in the systems table and export
the filtered data to pdf, but it appears that filtering doesn't work for
pdf export. I've created a card to fix this issue: https://issues.redhat.com/browse/RHINENG-17627